Change int to long int to accomodate INT_MIN

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  */
 int print_last_digit(int n)
 {
-	int absVal, d;
+	long int absVal, d;
 
 	if (n >= 0)
 		absVal = n;
